swiftvalidator/src/net/uglobal/swiftvalidator/swift/field/Field97B.java

package net.uglobal.swiftvalidator.swift.field;



/**
 * Field format : (qualifier)(dataSourceScheme)(accountTypeCode)(accountNumber)
 * 
 * AUTO GENERATED - DO NOT HAND MODIFY
 * Generated on Sun Dec 09 17:22:28 EST 2012
 * 
 */
public class Field97B
    extends net.uglobal.swiftvalidator.swift.field.Field
{


    public Field97B(String validateData, net.uglobal.swiftvalidator.xml.beans.Field xmlField) {
        super(validateData, xmlField);
    }

    @Override
    protected void parse() {
        addComponents(net.uglobal.swiftvalidator.swift.field.FieldComponentParseUtil.parse(validateData, getContentFormat()));
    }

    public String getQualifier() {
        return getComponent1();
    }

    public String getDataSourceScheme() {
        return getComponent2();
    }

    public String getAccountTypeCode() {
        return getComponent3();
    }

    public String getAccountNumber() {
        return getComponent4();
    }

    public String toString() {
        return (new StringBuilder()).append("[" ).append( "qualifier=" + getComponent1()).append( ", dataSourceScheme=" + getComponent2()).append( ", accountTypeCode=" + getComponent3()).append( ", accountNumber=" + getComponent4() ).append( "]").toString();
    }

}
